logging_service/requirements.txt
fastapi
uvicorn
hazelcast-python-client
